decouple chat session and router
