+++ docs/STATE.MD
@@ -123,4 +123,4 @@
 10/04/2026
 
 - [x] decouple tui by creaging a presentation layer pahse 1 and 2
-- []
+- [x] decouple chat session and router
